Unify the `load` interface
